+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Change Log
 - fixed deprecation warning about setting threads in daemon mode
 - fixed more threading module deprecation warnings
 - proxy now correctly exposes remote __len__, __iter__ and __getitem__ etc
+- improved type hint for expose()
 
 
 **Pyro 5.13.1**